requirements.txt
python-telegram-bot==20.7
openai==1.13.3
python-dotenv==1.0.1